Remove EventService and consolidate send test message logic into NotificationService
