+++ dashboards-notifications/models/interfaces.ts
@@ -9,24 +9,7 @@ import {
   CHANNEL_TYPE,
   ENCRYPTION_TYPE,
   NOTIFICATION_SOURCE,
-  SEVERITY_TYPE,
 } from '../public/utils/constants';
-
-export interface NotificationItem {
-  event_id: string;
-  created_time_ms: number;
-  last_updated_time_ms: number;
-  tenant?: string;
-  event_source: {
-    title: string;
-    reference_id: string;
-    feature: keyof typeof NOTIFICATION_SOURCE;
-    severity: keyof typeof SEVERITY_TYPE;
-    tags?: string[];
-  };
-  status_list: ChannelStatus[]; // could be multiple channels in a notification item
-  success: boolean; // calculated in the frontend based on status_list
-}
 
 export interface ChannelStatus {
   config_id: string;
