+++ dashboards-notifications/public/application.tsx
@@ -9,14 +9,13 @@ import { HashRouter as Router, Route } from 'react-router-dom';
 import { AppMountParameters, CoreStart } from '../../../src/core/public';
 import { CoreServicesContext } from './components/coreServices';
 import Main from './pages/Main';
-import { EventService, NotificationService } from './services';
+import { NotificationService } from './services';
 import { ServicesContext } from './services/services';
 
 export const renderApp = (coreStart: CoreStart, params: AppMountParameters) => {
   const http = coreStart.http;
   const notificationService = new NotificationService(http);
-  const eventService = new EventService(http);
-  const services = { notificationService, eventService };
+  const services = { notificationService };
 
   ReactDOM.render(
     <Router>
