+++ dashboards-notifications/public/services/index.ts
@@ -3,8 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import EventService from './EventService';
 import NotificationService from './NotificationService';
 import { ServicesConsumer, ServicesContext } from './services';
 
-export { ServicesConsumer, ServicesContext, NotificationService, EventService };
+export { ServicesConsumer, ServicesContext, NotificationService };
